Add I/O logic to deal with data file
"assignment.pdf" is the file containing the
manual from which our whole source code is based.

According to instructions given in assignment.pdf,
that is, informations related to ADT, data file
formats, etc., the code can basically read a file
and insert its data inside the backpack variable,
which is composed by items that have weigh and
value as their attributes, each.

Although the project structure is still very basic,
we intend to automate some compilation tasks and
add some directories later on.